Implementando novas chamadas, serviços e ajustes

+++ src/main/java/br/com/delivery/v1/repository/EstadoRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface EstadoRepository {
 
     List<Estado> listar();
-    Estado buscar(Long id);
+    Estado buscar(Integer id);
     Estado salvar(Estado estado);
     void remover(Estado estado);
